Add thread/process safety when running under WSGI.
This resolved three issues encountered when running under WSGI where
race conditions were observed when executing simultaneous requests.

1) In the top level WPS initialization, only load configuration if it has not
previously been loaded.  The key here is that the config object is stored
globally.  Observed failures in reading from the config in one process if
another processes attempted to simultaneously load, thereby obliterating
or invalidating the existing config.
2) Issues with reading from and writing to the correct directory when storing
complex results from a (c)StringIO object to a file and/or then copying this
file to another file in the configured output path.  This is due to several
os.chdir() calls throughout the codebase which twiddle the current operating
directory.  To work around this, (c)StringIO objects now are serialized to a temp
file within the configured output pathi, and when setting output raw data if
the source file is already within the output path then it is no longer
copied.
3) In Execute, dirsToBeRemoved to changed from a class variable to an instance
variable so that it is no longer shared between instances.  Previously, due to
sharing, removal of directories was sometimes failing due to the dirsToBeRemoved
list being manipulated simultaneously in multiple processes.